refactor: UI 패키지 분리
